refactor(authoring): drop the unbuilt channel-manifest apparatus
`channelManifest.ts` carried a manifest, element fingerprints, bake-loss
reasons, baked-clip provenance and a channel-drift classification for a
triage feature that was never built. Of its fourteen exports exactly one,
`dedupeImportedClips`, had a caller; `ChannelDrift*` was referenced
nowhere at all. Provenance in particular was speculative twice over,
since the shipped implementation went into `bakedAnimationProvenance.ts`
instead.

The one live function moves to a module named after it, keeping the
canonicalizer it needs.

`resolveGltfAnimationChannels` is *not* dead and stays: nothing on the
import path calls it, but the Blender corpus regression suite does, and
that suite's golden values are real coverage. The actual problem there is
that `convertGltfAnimations` resolves inline instead of calling it — one
resolver tested and another one shipped. That is a behaviour-affecting
consolidation, not a deletion, so it is left for its own change and the
barrel now says so.
