+++ apps/vizij-authoring/src/animationImport/index.ts
@@ -1,4 +1,8 @@
-export * from "./channelManifest";
+export * from "./dedupeImportedClips";
+// Not on the import path: `convertGltfAnimations` resolves inline. Kept
+// because the Blender corpus regression suite tests it, which makes the real
+// issue a duplicated resolver rather than a dead module.
+export * from "./resolveGltfAnimationChannels";
 export * from "./convertGltfAnimations";
 export * from "./gltfAccessors";
 export * from "./gltfAnimationChannels";
@@ -7,4 +11,3 @@ export * from "./importGltfAnimations";
 export * from "./inputRangeFit";
 export * from "./propsRigTargetCatalog";
 export * from "./quaternionToEuler";
-export * from "./resolveGltfAnimationChannels";
